Various improvements to gender issues.

+++ include/permonst.h
@@ -53,7 +53,7 @@ struct permonst {
 		mc,                     /* (base) magic cancellation (against magical and supernatural touch attacks) */
 		mr;                     /* (base) magic resistance (against all other magic) */
     aligntyp maligntyp;         /* basic monster alignment */
-    unsigned short geno;        /* creation/geno mask value */
+    unsigned long geno;         /* creation/geno mask value */
     struct attack mattk[NATTK]; /* attacks matrix */
     unsigned short cwt,         /* weight of corpse */
         cnutrit;                /* its nutritional value */
